1.0/requirements.txt
Flask
flask-socketio
gevent
gunicorn[gevent]
python-dotenv
colorama
